WEB/[Simple KB]/src/requirements.txt
Flask==3.1.0
Requests==2.32.3
PyJWT
PyYAML
pathlib